LayoutTests/editing/spelling/spelling-with-punctuation-selection-expected.txt

Punctuation marks should be treated as whitespace: spelling should ignore them. To test manually, type 'wellcome!' and select 'wellcome!'. The test succeeds when 'wellcome' has spelling marker.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS internals.markerCountForNode(destination.childNodes[0], "spelling") became different from 0
PASS spellingMarkerRange.toString() is "wellcome"
PASS window.getSelection().toString() is ""
PASS window.getSelection().toString() is "wellcome!"
PASS spellingMarkerRange.toString() became "wellcome"
PASS successfullyParsed is true

TEST COMPLETE
wellcome!
